Operating model class

Operating model class definition.

Usage

om(
  ages,
  time,
  harvest_function = function(object, numbers, selectivity, pst, i) {
    
    ifelse(all(is.na(object@targets$harvest_rate)), 0, object@targets$harvest_rate[i])
 },
  ...
)

Arguments

ages

a numeric value giving the maximum age or a vector of ages

time

a numeric value giving the number of time steps or a vector of times

harvest_function

harvest rate function for use in projection

...

optional input arguments with default values: samples = 1, shape = 1, phi = 1

Slots

.Data

numbers generated by the model run as the sum of 1+ age classes.

ages

integer vector of ages assumed by operating model. Set to NA when a cohort aggregated model is assumed.

time

integer vector of times used for operating model projection or single value giving the number of time steps.

samples

integer value indicating number of samples from the input value distributions specified in pars.

settings

list of settings used to for reference point evaluation with shape and rp.

pars

list of input parameter distributions used by the operating model. See load_pars.

shape

numeric vector of shape values estimated or specified using shape.

harvest_rate

function containing the harvest rate function.

pst

list containing phi, rmax and value elements related to the PST threshold reference point.

targets

list containing capture, depletion and harvest_rate target reference points estimated using rp.

objectives

list containing probability values indicating whether management target has been reached (i.e., the realised objective values).

diagnostics

list containing realised model outputs capture, depletion and harvest_rate.

values

data frame of values sampled from the life-history input distributions during execution of the model.

seeds

random number seeds generated when the om is initialised and used throughout.